+++ lib/change_health/response/trading_partner/trading_partners_data.rb
@@ -2,11 +2,7 @@ module ChangeHealth
   module Response
     class TradingPartnersData < Array
       def initialize(trading_partners)
-        super(trading_partners.map {|trading_partner| ChangeHealth::Response::TradingPartnerData.new(trading_partner) })
-      end
-
-      def medical_eligibility_enabled
-        self.reject {|partner| nil == partner.medical_eligibility_service_id }
+        super(trading_partners.map { |trading_partner| ChangeHealth::Response::TradingPartnerData.new(trading_partner) })
       end
     end
   end